frontend/requirements.txt
asgiref==3.7.2
cfgv==3.4.0
cmdstanpy==1.2.1
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
Django==5.0.3
django-debug-toolbar==4.3.0
filelock==3.13.1
fonttools==4.49.0
holidays==0.44
identify==2.5.35
importlib_resources==6.1.2
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
nodeenv==1.8.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
pre-commit==3.6.2
psycopg2==2.9.9
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
stanio==0.3.0
tenacity==8.2.3
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
virtualenv==20.25.1
